requirements.txt
Django==1.6.2
beautifulsoup4==4.3.2
dj-database-url==0.2.1
django-annoying==0.7.6
django-geoposition==0.1.5
django-storages==1.1.5
gunicorn==0.15.0
icalendar==3.1
jsonfield==0.9.19
odict==1.5.1
ply==3.4
python-dateutil==2.1
pytz==2012g
requests==0.13.2
simplejson==2.6.2
six==1.2.0
slimit==0.7.4
stripe==1.9.8
wsgiref==0.1.2
